refactor(core): move things around a fix a TS problem in hot-server
