+++ packages/core/src/scripts/dev.ts
@@ -1,16 +1,72 @@
 import Argv from "minimist";
 import { ensureDir, emptyDir } from "fs-extra";
+import express from "express";
 import webpack from "webpack";
 import webpackDevMiddleware from "webpack-dev-middleware";
 import webpackHotMiddleware from "webpack-hot-middleware";
-import webpackHotServerMiddleware from "./hot-server";
+import createServer from "./utils/create-server";
+import webpackHotServerMiddleware from "./utils/hot-server";
 import getConfig from "../config";
 import { Mode } from "../types";
-import { createApp } from "./utils";
 
 const buildDir = "build";
 const analyzeDir = "analyze";
 const argv = Argv(process.argv.slice(2));
+
+const createApp = async ({
+  mode,
+  port,
+  isHttps,
+  es5
+}: {
+  mode: Mode;
+  port: number;
+  isHttps: boolean;
+  es5: boolean;
+}): Promise<{
+  app: express.Express;
+  done: (compiler: webpack.MultiCompiler) => void;
+}> => {
+  // Create the server.
+  const app = express();
+  // Create a function to start listening after webpack has finished.
+  const server = await createServer({ app, isHttps });
+  let clientFinished = false;
+  let serverFinished = false;
+  const start = () => {
+    if (clientFinished && serverFinished) {
+      server.listen(port, () => {
+        console.log(
+          `\n\nSERVER STARTED -- Listening @ ${
+            isHttps ? "https" : "http"
+          }://localhost:${port}\n  - mode: ${mode}\n  - client: ${
+            es5 ? "es5" : "esModules"
+          }`
+        );
+      });
+    }
+  };
+  const done = (compiler: webpack.MultiCompiler) => {
+    compiler.compilers[0].hooks.done.tapAsync(
+      "frontity-dev-server",
+      (_, cb) => {
+        clientFinished = true;
+        start();
+        cb();
+      }
+    );
+    compiler.compilers[1].hooks.done.tapAsync(
+      "frontity-dev-server",
+      (_, cb) => {
+        serverFinished = true;
+        start();
+        cb();
+      }
+    );
+  };
+
+  return { app, done };
+};
 
 const dev = async ({
   isHttps,
