--- packages/core/src/scripts/utils.ts
@@ -1,78 +1,0 @@
-import { readFile } from "fs-extra";
-import express from "express";
-import { Mode } from "../types";
-import * as http from "http";
-import * as https from "https";
-import webpack = require("webpack");
-
-// Create HTTP or HTTPS server using a self-signed certificate.
-export const createServer = async ({
-  app,
-  isHttps = true
-}: {
-  app: express.Express;
-  isHttps: boolean;
-}): Promise<http.Server | https.Server> => {
-  if (isHttps) {
-    const key = await readFile("src/certs/localhost.key");
-    const cert = await readFile("src/certs/localhost.cert");
-    const options: https.ServerOptions = { key, cert };
-    return https.createServer(options, app);
-  }
-  return http.createServer(app);
-};
-
-export const createApp = async ({
-  mode,
-  port,
-  isHttps,
-  es5
-}: {
-  mode: Mode;
-  port: number;
-  isHttps: boolean;
-  es5: boolean;
-}): Promise<{
-  app: express.Express;
-  done: (compiler: webpack.MultiCompiler) => void;
-}> => {
-  // Create the server.
-  const app = express();
-  // Create a function to start listening after webpack has finished.
-  const server = await createServer({ app, isHttps });
-  let clientFinished = false;
-  let serverFinished = false;
-  const start = () => {
-    if (clientFinished && serverFinished) {
-      server.listen(port, () => {
-        console.log(
-          `\n\nSERVER STARTED -- Listening @ ${
-            isHttps ? "https" : "http"
-          }://localhost:${port}\n  - mode: ${mode}\n  - client: ${
-            es5 ? "es5" : "esModules"
-          }`
-        );
-      });
-    }
-  };
-  const done = (compiler: webpack.MultiCompiler) => {
-    compiler.compilers[0].hooks.done.tapAsync(
-      "frontity-dev-server",
-      (_, cb) => {
-        clientFinished = true;
-        start();
-        cb();
-      }
-    );
-    compiler.compilers[1].hooks.done.tapAsync(
-      "frontity-dev-server",
-      (_, cb) => {
-        serverFinished = true;
-        start();
-        cb();
-      }
-    );
-  };
-
-  return { app, done };
-};
